+++ src/main/java/com/yuuko/core/commands/developer/commands/ReloadApiCommand.java
@@ -3,18 +3,21 @@ package com.yuuko.core.commands.developer.commands;
 import com.yuuko.core.Configuration;
 import com.yuuko.core.commands.Command;
 import com.yuuko.core.commands.developer.DeveloperModule;
+import com.yuuko.core.utilities.MessageHandler;
+import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 
 public class ReloadApiCommand extends Command {
 
     public ReloadApiCommand() {
-        super("reapi", DeveloperModule.class, 1, new String[]{"-reapi"}, false, null);
+        super("reapi", DeveloperModule.class, 0, new String[]{"-reapi"}, false, null);
     }
 
     @Override
     public void onCommand(MessageReceivedEvent e, String[] command) {
         try {
-            Configuration.loadApi();
+            EmbedBuilder embed = new EmbedBuilder().setTitle("Successfully reloaded " + Configuration.loadApi() + " API Keys.");
+            MessageHandler.sendMessage(e, embed.build());
         } catch(Exception ex) {
             log.error("An error occurred while running the {} class, message: {}", this, ex.getMessage(), ex);
         }