+++ src/main/java/com/yuuko/core/commands/world/WorldModule.java
@@ -3,10 +3,7 @@ package com.yuuko.core.commands.world;
 import com.yuuko.core.CommandExecutor;
 import com.yuuko.core.commands.Command;
 import com.yuuko.core.commands.Module;
-import com.yuuko.core.commands.world.commands.CountdownCommand;
-import com.yuuko.core.commands.world.commands.LondonUndergroundCommand;
-import com.yuuko.core.commands.world.commands.TescoCommand;
-import com.yuuko.core.commands.world.commands.WeatherCommand;
+import com.yuuko.core.commands.world.commands.*;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 
 public class WorldModule extends Module {
@@ -16,7 +13,8 @@ public class WorldModule extends Module {
                 new LondonUndergroundCommand(),
                 new WeatherCommand(),
                 new TescoCommand(),
-                new CountdownCommand()
+                new CountdownCommand(),
+                new NationalGeographicCommand()
         });
 
         new CommandExecutor(e,this, command);